Fix automatic click on preview button causing infinite refresh cycle. Fix webview url loads loading in external browser.
